fixed two warnings in WebSocket[Impl]

+++ Net/src/WebSocket.cpp
@@ -1,7 +1,7 @@
 //
 // WebSocket.cpp
 //
-// $Id: //poco/1.4/Net/src/WebSocket.cpp#3 $
+// $Id: //poco/1.4/Net/src/WebSocket.cpp#4 $
 //
 // Library: Net
 // Package: WebSocket
@@ -107,7 +107,7 @@ void WebSocket::shutdown(Poco::UInt16 statusCode, const std::string& statusMessa
 	Poco::BinaryWriter writer(ostr, Poco::BinaryWriter::NETWORK_BYTE_ORDER);
 	writer << statusCode;
 	writer.writeRaw(statusMessage);
-	sendFrame(buffer.begin(), ostr.charsWritten(), FRAME_FLAG_FIN | FRAME_OP_CLOSE);
+	sendFrame(buffer.begin(), static_cast<int>(ostr.charsWritten()), FRAME_FLAG_FIN | FRAME_OP_CLOSE);
 }
